requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
yt-dlp>=2023.12.30
requests>=2.31.0
PyNaCl>=1.5.0
